+++ src/modules/graphics/graphics.c
@@ -29,10 +29,12 @@
 
 #define MAX_PIPELINES 8192
 #define MAX_TALLIES 255
+#define MATERIAL_BLOCK_SIZE 256
 #define TRANSFORM_STACK_SIZE 16
 #define PIPELINE_STACK_SIZE 8
 #define MAX_SHADER_RESOURCES 32
 #define MAX_CUSTOM_ATTRIBUTES 10
+#define MATERIAL_STRIDE ALIGN(sizeof(MaterialData), state.limits.uniformBufferAlign)
 #define FLOAT_BITS(f) ((union { float f; uint32_t u; }) { f }).u
 
 typedef struct {
@@ -177,25 +179,40 @@ struct Shader {
 };
 
 typedef struct {
+  float numbers[NUMBER_COUNT];
+  float padding;
+  float colors[COLOR_COUNT][4];
+  float quad[4];
+  float sdfRange[2];
+} MaterialData;
+
+typedef struct {
+  uint32_t next;
+  uint32_t tick;
+} MaterialNode;
+
+typedef struct {
   void* next;
-  void* pointer;
-  Material* materials;
-  gpu_buffer* buffer;
-  gpu_bundle_pool* bundlePool;
-  gpu_bundle* bundles;
   uint32_t head;
   uint32_t tail;
+  MaterialNode nodes[MATERIAL_BLOCK_SIZE];
+  gpu_buffer* buffer;
+  void* bufferPointer;
+  gpu_bundle_pool* bundlePool;
+  gpu_bundle* bundles;
 } MaterialBlock;
 
 struct Material {
   atomic_uint ref;
-  uint32_t next;
-  uint32_t tick;
   uint32_t index;
-  MaterialInfo info;
-  gpu_bundle* bundle;
   MaterialBlock* block;
-  bool hasWritableTexture;
+  MaterialData data;
+  Texture* textures[TEXTURE_COUNT];
+  gpu_binding bindings[1 + TEXTURE_COUNT];
+  gpu_bundle* bundle;
+  MaterialData* pointer;
+  uint32_t bundleTick;
+  uint32_t bufferTick;
 };
 
 typedef struct {
@@ -792,15 +809,7 @@ bool lovrGraphicsInit(GraphicsConfig* config) {
 
   // Default Material
 
-  state.defaultMaterial = lovrMaterialCreate(&(MaterialInfo) {
-    .data.color = { 1.f, 1.f, 1.f, 1.f },
-    .data.uvScale = { 1.f, 1.f },
-    .data.metalness = 0.f,
-    .data.roughness = 1.f,
-    .data.normalScale = 1.f,
-    .texture = state.defaultTexture
-  });
-
+  state.defaultMaterial = lovrMaterialCreate(NULL);
   if (!state.defaultMaterial) goto fail;
 
   // Default Samplers
@@ -918,7 +927,6 @@ void lovrGraphicsDestroy(void) {
     MaterialBlock* next = block->next;
     gpu_bundle_pool_destroy(block->bundlePool);
     gpu_buffer_destroy(block->buffer);
-    lovrFree(block->materials);
     lovrFree(block->bundlePool);
     lovrFree(block->bundles);
     lovrFree(block->buffer);
@@ -3100,11 +3108,7 @@ void lovrTextureSetSampler(Texture* texture, Sampler* sampler) {
 
 Material* lovrTextureToMaterial(Texture* texture) {
   if (!texture->material) {
-    texture->material = lovrMaterialCreate(&(MaterialInfo) {
-      .data.color = { 1.f, 1.f, 1.f, 1.f },
-      .data.uvScale = { 1.f, 1.f },
-      .texture = texture
-    });
+    texture->material = lovrMaterialCreate(texture);
 
     if (!texture->material) {
       return NULL;
@@ -3113,7 +3117,7 @@ Material* lovrTextureToMaterial(Texture* texture) {
     // Since the Material refcounts the Texture, this creates a cycle.  Release the texture to make
     // sure this is a weak relationship (the automaterial does not keep the texture refcounted).
     lovrRelease(texture, lovrTextureDestroy);
-    texture->material->info.texture = NULL;
+    texture->material->textures[TEXTURE_COLOR] = NULL;
   }
 
   return texture->material;
@@ -4116,164 +4120,272 @@ const DataField* lovrShaderGetBufferFormat(Shader* shader, const char* name, uin
 
 // Material
 
-Material* lovrMaterialCreate(const MaterialInfo* info) {
-  Texture* textures[] = {
-    info->texture,
-    info->glowTexture,
-    info->metalnessTexture,
-    info->roughnessTexture,
-    info->clearcoatTexture,
-    info->occlusionTexture,
-    info->normalTexture
-  };
-
-  for (uint32_t i = 0; i < COUNTOF(textures); i++) {
-    if (!textures[i]) continue;
-    lovrCheck(i == 0 || textures[i]->info.type == TEXTURE_2D, "Material textures must be 2D");
-    lovrCheck(textures[i]->info.samples == 1, "Material textures can not be multisampled");
-    lovrCheck(textures[i]->info.usage & TEXTURE_SAMPLE, "Textures must be created with the 'sample' usage to use them in Materials");
-  }
-
-  // TODO this should be more fine-grained
+static bool lovrMaterialAllocate(Material* material) {
   mtx_lock(&state.lock);
 
   MaterialBlock* block = NULL;
-
-  for (MaterialBlock* node = state.materials; node != NULL; node = node->next) {
-    if (node->head != ~0u && gpu_is_complete(node->materials[node->head].tick)) {
-      block = node;
+  for (MaterialBlock* b = state.materials; b != NULL; b = b->next) {
+    if (b->head != ~0u && gpu_is_complete(b->nodes[b->head].tick)) {
+      block = b;
       break;
     }
   }
 
   if (!block) {
-    const uint32_t count = 256;
-    block = lovrMalloc(sizeof(*block));
-    block->materials = lovrMalloc(count * sizeof(Material));
+    block = lovrMalloc(sizeof(MaterialBlock));
     block->buffer = lovrMalloc(gpu_sizeof_buffer());
     block->bundlePool = lovrMalloc(gpu_sizeof_bundle_pool());
-    block->bundles = lovrMalloc(count * gpu_sizeof_bundle());
+    block->bundles = lovrMalloc(MATERIAL_BLOCK_SIZE * gpu_sizeof_bundle());
 
-    for (uint32_t i = 0; i < count; i++) {
-      block->materials[i] = (Material) {
-        .index = i,
-        .next = i + 1,
-        .block = block,
-        .bundle = (gpu_bundle*) ((char*) block->bundles + i * gpu_sizeof_bundle())
-      };
+    for (uint32_t i = 0; i < MATERIAL_BLOCK_SIZE; i++) {
+      block->nodes[i].next = i + 1;
+      block->nodes[i].tick = 0;
     }
 
-    block->materials[count - 1].next = ~0u;
-    block->tail = count - 1;
     block->head = 0;
+    block->tail = MATERIAL_BLOCK_SIZE - 1;
+    block->nodes[block->tail].next = ~0u;
 
     gpu_bundle_pool_info poolInfo = {
       .bundles = block->bundles,
       .layout = state.materialLayout->gpu,
-      .count = count
+      .count = MATERIAL_BLOCK_SIZE
     };
 
     if (!gpu_bundle_pool_init(block->bundlePool, &poolInfo)) {
-      lovrFree(block->materials);
       lovrFree(block->buffer);
       lovrFree(block->bundlePool);
       lovrFree(block->bundles);
       mtx_unlock(&state.lock);
-      return NULL;
+      return false;
     }
 
     gpu_buffer_info bufferInfo = {
       .type = GPU_BUFFER_STATIC,
-      .size = count * (uint32_t) ALIGN(sizeof(MaterialData), state.limits.uniformBufferAlign),
-      .pointer = &block->pointer
+      .size = MATERIAL_BLOCK_SIZE * MATERIAL_STRIDE,
+      .pointer = &block->bufferPointer
     };
 
     if (!gpu_buffer_init(block->buffer, &bufferInfo)) {
-      lovrFree(block->materials);
       lovrFree(block->buffer);
       lovrFree(block->bundlePool);
       lovrFree(block->bundles);
       gpu_bundle_pool_destroy(block->bundlePool);
       mtx_unlock(&state.lock);
-      return NULL;
+      return false;
     }
 
     block->next = state.materials;
     state.materials = block;
   }
 
-  Material* material = &block->materials[block->head];
-  material->ref = 1;
-  material->info = *info;
+  material->block = block;
+  material->index = block->head;
+  material->bundle = (gpu_bundle*) ((char*) block->bundles + block->head * gpu_sizeof_bundle());
+  material->bundleTick = state.tick;
 
-  MaterialData* data;
-  uint32_t stride = ALIGN(sizeof(MaterialData), state.limits.uniformBufferAlign);
+  material->bindings[0] = (gpu_binding) {
+    .number = 0,
+    .type = GPU_SLOT_UNIFORM_BUFFER,
+    .buffer.object = block->buffer,
+    .buffer.offset = material->index * MATERIAL_STRIDE,
+    .buffer.extent = MATERIAL_STRIDE
+  };
 
-  if (block->pointer) {
-    data = (MaterialData*) ((char*) block->pointer + material->index * stride);
-  } else {
-    BufferView staging = getBuffer(GPU_BUFFER_UPLOAD, sizeof(MaterialData), 4);
-    if (!staging.buffer) return mtx_unlock(&state.lock), NULL;
-
-    gpu_copy_buffers(state.stream, staging.buffer, block->buffer, staging.offset, stride * material->index, sizeof(MaterialData));
-    state.barrier.prev |= GPU_PHASE_COPY;
-    state.barrier.next |= GPU_PHASE_SHADER_VERTEX | GPU_PHASE_SHADER_FRAGMENT;
-    state.barrier.flush |= GPU_CACHE_TRANSFER_WRITE;
-    state.barrier.clear |= GPU_CACHE_UNIFORM;
-    data = staging.pointer;
+  if (block->bufferPointer) {
+    material->pointer = (MaterialData*) ((char*) block->bufferPointer + material->index * MATERIAL_STRIDE);
+    material->bufferTick = state.tick;
   }
 
-  memcpy(data, info, sizeof(MaterialData));
+  block->head = block->nodes[block->head].next;
+  block->nodes[material->index].next = ~0u;
 
-  gpu_buffer_binding buffer = {
-    .object = block->buffer,
-    .offset = material->index * stride,
-    .extent = stride
-  };
-
-  gpu_binding bindings[8] = {
-    { 0, GPU_SLOT_UNIFORM_BUFFER, .buffer = buffer }
-  };
-
-  for (uint32_t i = 0; i < COUNTOF(textures); i++) {
-    Texture* texture = textures[i] ? textures[i] : state.defaultTexture;
-    bindings[i + 1] = (gpu_binding) { i + 1, GPU_SLOT_SAMPLED_TEXTURE, .texture.object = texture->sampleViewFloat };
-    material->hasWritableTexture |= texture->info.usage != TEXTURE_SAMPLE;
-    lovrRetain(textures[i]);
-  }
-
-  gpu_bundle_info bundleInfo = {
-    .layout = state.materialLayout->gpu,
-    .bindings = bindings,
-    .count = COUNTOF(bindings)
-  };
-
-  gpu_bundle_write(&material->bundle, &bundleInfo, 1);
-
-  block->head = material->next;
-  material->next = ~0u;
   mtx_unlock(&state.lock);
+
+  return true;
+}
+
+static void lovrMaterialRecycle(Material* material) {
+  if (!material->block) return;
+  mtx_lock(&state.lock);
+  material->block->nodes[material->index].tick = state.tick;
+  material->block->tail = material->index;
+  if (material->block->head == ~0u) material->block->head = material->block->tail;
+  material->block = NULL;
+  material->index = ~0u;
+  material->bundle = NULL;
+  mtx_unlock(&state.lock);
+}
+
+static bool lovrMaterialUpload(Material* material) {
+  if (material->bufferTick == state.tick) {
+    return true;
+  }
+
+  mtx_lock(&state.lock);
+  BufferView staging = getBuffer(GPU_BUFFER_STREAM, sizeof(MaterialData), 4);
+  if (!staging.buffer) return mtx_unlock(&state.lock), NULL;
+  gpu_copy_buffers(state.stream, staging.buffer, material->block->buffer, staging.offset, MATERIAL_STRIDE * material->index, sizeof(MaterialData));
+  state.barrier.prev |= GPU_PHASE_COPY;
+  state.barrier.next |= GPU_PHASE_SHADER_VERTEX | GPU_PHASE_SHADER_FRAGMENT;
+  state.barrier.flush |= GPU_CACHE_TRANSFER_WRITE;
+  state.barrier.clear |= GPU_CACHE_UNIFORM;
+  material->bufferTick = state.tick;
+  material->pointer = staging.pointer;
+  memcpy(material->pointer, &material->data, sizeof(MaterialData));
+  mtx_unlock(&state.lock);
+
+  return true;
+}
+
+Material* lovrMaterialCreate(Texture* texture) {
+  Material* material = lovrCalloc(sizeof(Material));
+  material->ref = 1;
+
+  material->data = (MaterialData) {
+    .colors[COLOR_BASE] = { 1.f, 1.f, 1.f, 1.f },
+    .quad = { 0.f, 0.f, 1.f, 1.f }
+  };
+
+  for (uint32_t i = 0; i < TEXTURE_COUNT; i++) {
+    Texture* t = state.defaultTexture;
+    material->textures[i] = t;
+    material->bindings[i + 1] = (gpu_binding) {
+      .number = i + 1,
+      .type = GPU_SLOT_SAMPLED_TEXTURE,
+      .texture.object = t->sampleViewFloat
+    };
+    lovrRetain(t);
+  }
+
+  if (texture) {
+    // This will take care of allocating the material and writing all the bindings
+    if (!lovrMaterialSetTexture(material, TEXTURE_COLOR, texture)) {
+      lovrFree(material);
+      return NULL;
+    }
+  } else {
+    if (!lovrMaterialAllocate(material)) {
+      lovrFree(material);
+      return NULL;
+    }
+
+    gpu_bundle_info bundleInfo = {
+      .layout = state.materialLayout->gpu,
+      .bindings = material->bindings,
+      .count = COUNTOF(material->bindings)
+    };
+
+    gpu_bundle_write(&material->bundle, &bundleInfo, 1);
+  }
+
+  if (material->pointer) {
+    memcpy(material->pointer, &material->data, sizeof(MaterialData));
+  } else {
+    lovrMaterialUpload(material);
+  }
+
   return material;
 }
 
 void lovrMaterialDestroy(void* ref) {
   Material* material = ref;
-  mtx_lock(&state.lock);
-  material->tick = state.tick;
-  material->block->tail = material->index;
-  if (material->block->head == ~0u) material->block->head = material->block->tail;
-  mtx_unlock(&state.lock);
-  lovrRelease(material->info.texture, lovrTextureDestroy);
-  lovrRelease(material->info.glowTexture, lovrTextureDestroy);
-  lovrRelease(material->info.metalnessTexture, lovrTextureDestroy);
-  lovrRelease(material->info.roughnessTexture, lovrTextureDestroy);
-  lovrRelease(material->info.clearcoatTexture, lovrTextureDestroy);
-  lovrRelease(material->info.occlusionTexture, lovrTextureDestroy);
-  lovrRelease(material->info.normalTexture, lovrTextureDestroy);
+  lovrMaterialRecycle(material);
+  for (uint32_t i = 0; i < TEXTURE_COUNT; i++) {
+    lovrRelease(material->textures[i], lovrTextureDestroy);
+  }
+  lovrFree(material);
 }
 
-const MaterialInfo* lovrMaterialGetInfo(Material* material) {
-  return &material->info;
+float lovrMaterialGetNumber(Material* material, MaterialNumber key) {
+  return material->data.numbers[key];
+}
+
+bool lovrMaterialSetNumber(Material* material, MaterialNumber key, float number) {
+  if (!lovrMaterialUpload(material)) return false;
+  material->data.numbers[key] = number;
+  material->pointer->numbers[key] = number;
+  return true;
+}
+
+const float* lovrMaterialGetColor(Material* material, MaterialColor key) {
+  return material->data.colors[key];
+}
+
+bool lovrMaterialSetColor(Material* material, MaterialColor key, float color[4]) {
+  if (!lovrMaterialUpload(material)) return false;
+  memcpy(material->data.colors[key], color, 4 * sizeof(float));
+  material->pointer->colors[key][0] = lovrMathGammaToLinear(color[0]);
+  material->pointer->colors[key][1] = lovrMathGammaToLinear(color[1]);
+  material->pointer->colors[key][2] = lovrMathGammaToLinear(color[2]);
+  material->pointer->colors[key][3] = color[3];
+  return true;
+}
+
+Texture* lovrMaterialGetTexture(Material* material, MaterialTexture key) {
+  return material->textures[key];
+}
+
+bool lovrMaterialSetTexture(Material* material, MaterialTexture key, Texture* texture) {
+  if (texture == material->textures[key]) return true;
+
+  lovrCheck(key == TEXTURE_COLOR || texture->info.type == TEXTURE_2D, "Material textures must be 2D");
+  lovrCheck(texture->info.samples == 1, "Material textures can not be multisampled");
+  lovrCheck(texture->info.usage & TEXTURE_SAMPLE, "Textures must be created with the 'sample' usage to use them in Materials");
+
+  lovrRelease(material->textures[key], lovrTextureDestroy);
+  material->textures[key] = texture;
+  material->bindings[key + 1].texture.object = texture->sampleViewFloat;
+  lovrRetain(texture);
+
+  gpu_bundle_info bundleInfo = { .layout = state.materialLayout->gpu };
+
+  if (material->bundleTick == state.tick) {
+    bundleInfo.bindings = &material->bindings[key + 1];
+    bundleInfo.count = 1;
+  } else {
+    lovrMaterialRecycle(material);
+
+    if (!lovrMaterialAllocate(material)) {
+      return false;
+    }
+
+    bundleInfo.bindings = material->bindings;
+    bundleInfo.count = COUNTOF(material->bindings);
+  }
+
+  gpu_bundle_write(&material->bundle, &bundleInfo, 1);
+
+  return true;
+}
+
+void lovrMaterialGetQuad(Material* material, float* ox, float* oy, float* sx, float* sy) {
+  *ox = material->data.quad[0];
+  *oy = material->data.quad[1];
+  *sx = material->data.quad[2];
+  *sy = material->data.quad[3];
+}
+
+bool lovrMaterialSetQuad(Material* material, float ox, float oy, float sx, float sy) {
+  if (!lovrMaterialUpload(material)) return false;
+  material->data.quad[0] = ox;
+  material->data.quad[1] = oy;
+  material->data.quad[2] = sx;
+  material->data.quad[3] = sy;
+  material->pointer->quad[0] = ox;
+  material->pointer->quad[1] = oy;
+  material->pointer->quad[2] = sx;
+  material->pointer->quad[3] = sy;
+  return true;
+}
+
+static bool lovrMaterialSetSDFRange(Material* material, float x, float y) {
+  if (!lovrMaterialUpload(material)) return false;
+  material->data.sdfRange[0] = x;
+  material->data.sdfRange[1] = y;
+  material->pointer->sdfRange[0] = x;
+  material->pointer->sdfRange[1] = y;
+  return true;
 }
 
 // Font
@@ -4344,11 +4456,7 @@ Font* lovrFontCreate(const FontInfo* info) {
       return NULL;
     }
 
-    font->material = lovrMaterialCreate(&(MaterialInfo) {
-      .data.color = { 1.f, 1.f, 1.f, 1.f },
-      .data.uvScale = { 1.f, 1.f },
-      .texture = font->atlas
-    });
+    font->material = lovrMaterialCreate(font->atlas);
 
     if (!font->material) {
       lovrFontDestroy(font);
@@ -4514,12 +4622,8 @@ static Glyph* lovrFontGetGlyph(Font* font, uint32_t codepoint, bool* resized) {
       return NULL;
     }
 
-    Material* material = lovrMaterialCreate(&(MaterialInfo) {
-      .data.color = { 1.f, 1.f, 1.f, 1.f },
-      .data.uvScale = { 1.f, 1.f },
-      .data.sdfRange = { font->info.spread / newWidth, font->info.spread / newHeight },
-      .texture = atlas
-    });
+    Material* material = lovrMaterialCreate(atlas);
+    lovrMaterialSetSDFRange(material, font->info.spread / newWidth, font->info.spread / newHeight);
 
     if (!material) {
       lovrTextureDestroy(atlas);
@@ -5330,58 +5434,65 @@ Model* lovrModelCreate(const ModelInfo* info) {
   ModelData* data = info->data;
   ModelMetadata* meta = &model->meta;
 
-  // Materials and Textures
-  if (info->materials) {
-    model->textures = lovrCalloc(meta->imageCount * sizeof(Texture*));
-    model->materials = lovrMalloc(meta->materialCount * sizeof(Material*));
-    for (uint32_t i = 0; i < meta->materialCount; i++) {
-      MaterialInfo material;
-      ModelMaterial* properties = &meta->materials[i];
-      memcpy(&material.data, properties, sizeof(MaterialData));
+  // Materials
 
-      struct { uint32_t index; Texture** texture; } textures[] = {
-        { properties->texture, &material.texture },
-        { properties->glowTexture, &material.glowTexture },
-        { properties->metalnessTexture, &material.metalnessTexture },
-        { properties->roughnessTexture, &material.roughnessTexture },
-        { properties->clearcoatTexture, &material.clearcoatTexture },
-        { properties->occlusionTexture, &material.occlusionTexture },
-        { properties->normalTexture, &material.normalTexture }
+  model->textures = lovrCalloc(meta->imageCount * sizeof(Texture*));
+  model->materials = lovrCalloc(meta->materialCount * sizeof(Material*));
+
+  if (info->materials) {
+    for (uint32_t i = 0; i < meta->materialCount; i++) {
+      Material* material = lovrMaterialCreate(NULL);
+      lovrAssertGoto(fail, material, "Failed to create model material: %s", lovrGetError());
+      model->materials[i] = material;
+
+      ModelMaterial* properties = &meta->materials[i];
+      lovrMaterialSetNumber(material, NUMBER_METALNESS, properties->metalness);
+      lovrMaterialSetNumber(material, NUMBER_ROUGHNESS, properties->roughness);
+      lovrMaterialSetNumber(material, NUMBER_CLEARCOAT, properties->clearcoat);
+      lovrMaterialSetNumber(material, NUMBER_CLEARCOAT_ROUGHNESS, properties->clearcoatRoughness);
+      lovrMaterialSetNumber(material, NUMBER_OCCLUSION_STRENGTH, properties->occlusionStrength);
+      lovrMaterialSetNumber(material, NUMBER_NORMAL_SCALE, properties->normalScale);
+      lovrMaterialSetNumber(material, NUMBER_ALPHA_CUTOFF, properties->alphaCutoff);
+      lovrMaterialSetColor(material, COLOR_BASE, properties->color);
+      lovrMaterialSetColor(material, COLOR_GLOW, properties->glow);
+      lovrMaterialSetQuad(material, properties->uvShift[0], properties->uvShift[1], properties->uvScale[0], properties->uvScale[1]);
+
+      uint32_t textures[] = {
+        [TEXTURE_COLOR] = properties->texture,
+        [TEXTURE_GLOW] = properties->glowTexture,
+        [TEXTURE_METALNESS] = properties->metalnessTexture,
+        [TEXTURE_ROUGHNESS] = properties->roughnessTexture,
+        [TEXTURE_CLEARCOAT] = properties->clearcoatTexture,
+        [TEXTURE_OCCLUSION] = properties->occlusionTexture,
+        [TEXTURE_NORMAL] = properties->normalTexture
       };
 
       for (uint32_t t = 0; t < COUNTOF(textures); t++) {
-        uint32_t index = textures[t].index;
-        Texture** texture = textures[t].texture;
+        uint32_t index = textures[t];
+        if (index == ~0u) continue;
 
-        if (index == ~0u) {
-          *texture = NULL;
-        } else {
-          if (!model->textures[index]) {
-            Image* image = data->images[index];
+        if (!model->textures[index]) {
+          Image* image = data->images[index];
 
-            model->textures[index] = lovrTextureCreate(&(TextureInfo) {
-              .type = TEXTURE_2D,
-              .usage = TEXTURE_SAMPLE,
-              .format = lovrImageGetFormat(image),
-              .width = lovrImageGetWidth(image, 0),
-              .height = lovrImageGetHeight(image, 0),
-              .layers = 1,
-              .mipmaps = info->mipmaps || lovrImageGetLevelCount(image) > 1 ? ~0u : 1,
-              .samples = 1,
-              .srgb = texture == &material.texture || texture == &material.glowTexture,
-              .images = &image,
-              .imageCount = 1
-            });
+          model->textures[index] = lovrTextureCreate(&(TextureInfo) {
+            .type = TEXTURE_2D,
+            .usage = TEXTURE_SAMPLE,
+            .format = lovrImageGetFormat(image),
+            .width = lovrImageGetWidth(image, 0),
+            .height = lovrImageGetHeight(image, 0),
+            .layers = 1,
+            .mipmaps = info->mipmaps || lovrImageGetLevelCount(image) > 1 ? ~0u : 1,
+            .samples = 1,
+            .srgb = t == TEXTURE_COLOR || t == TEXTURE_GLOW,
+            .images = &image,
+            .imageCount = 1
+          });
 
-            if (!model->textures[index]) goto fail;
-          }
-
-          *texture = model->textures[index];
+          if (!model->textures[index]) goto fail;
         }
-      }
 
-      model->materials[i] = lovrMaterialCreate(&material);
-      lovrAssertGoto(fail, model->materials[i], "Failed to create model material: %s", lovrGetError());
+        lovrMaterialSetTexture(material, t, model->textures[index]);
+      }
     }
   }
 
@@ -5504,8 +5615,18 @@ Model* lovrModelClone(Model* parent) {
 
   ModelMetadata* meta = &model->meta;
 
-  model->textures = parent->textures;
-  model->materials = parent->materials;
+  model->textures = lovrMalloc(meta->imageCount * sizeof(Texture*));
+  model->materials = lovrMalloc(meta->materialCount * sizeof(Material*));
+  memcpy(model->textures, parent->textures, meta->imageCount * sizeof(Texture*));
+  memcpy(model->materials, parent->materials, meta->materialCount * sizeof(Material*));
+
+  for (uint32_t i = 0; i < meta->imageCount; i++) {
+    lovrRetain(model->textures[i]);
+  }
+
+  for (uint32_t i = 0; i < meta->materialCount; i++) {
+    lovrRetain(model->materials[i]);
+  }
 
   model->rawVertexBuffer = parent->rawVertexBuffer;
   model->indexBuffer = parent->indexBuffer;
@@ -5571,16 +5692,14 @@ void lovrModelDestroy(void* ref) {
     return;
   }
   ModelMetadata* meta = &model->meta;
-  if (model->materials) {
-    for (uint32_t i = 0; i < meta->materialCount; i++) {
-      lovrRelease(model->materials[i], lovrMaterialDestroy);
-    }
-    for (uint32_t i = 0; i < meta->imageCount; i++) {
-      lovrRelease(model->textures[i], lovrTextureDestroy);
-    }
-    lovrFree(model->materials);
-    lovrFree(model->textures);
+  for (uint32_t i = 0; i < meta->materialCount; i++) {
+    lovrRelease(model->materials[i], lovrMaterialDestroy);
   }
+  for (uint32_t i = 0; i < meta->imageCount; i++) {
+    lovrRelease(model->textures[i], lovrTextureDestroy);
+  }
+  lovrFree(model->materials);
+  lovrFree(model->textures);
   if (model->meshes) {
     for (uint32_t i = 0; i < meta->meshCount; i++) {
       lovrRelease(model->meshes[i], lovrMeshDestroy);
@@ -5833,7 +5952,7 @@ Mesh* lovrModelGetMesh(Model* model, uint32_t index) {
       default: lovrUnreachable();
     }
 
-    if (model->materials && part->material != ~0u) {
+    if (part->material != ~0u) {
       lovrMeshSetMaterial(mesh, model->materials[part->material]);
     }
 
@@ -5853,10 +5972,20 @@ Texture* lovrModelGetTexture(Model* model, uint32_t index) {
   return model->textures[index];
 }
 
-Material* lovrModelGetMaterial(Model* model, uint32_t index) {
+bool lovrModelGetMaterial(Model* model, uint32_t index, Material** material) {
   uint32_t count = model->meta.materialCount;
   lovrCheck(index < count, "Invalid material index '%d' (Model has %d material%s)", index + 1, count, count == 1 ? "" : "s");
-  return model->materials[index];
+  *material = model->materials[index];
+  return true;
+}
+
+bool lovrModelSetMaterial(Model* model, uint32_t index, Material* material) {
+  uint32_t count = model->meta.materialCount;
+  lovrCheck(index < count, "Invalid material index '%d' (Model has %d material%s)", index + 1, count, count == 1 ? "" : "s");
+  lovrRelease(model->materials[index], lovrMaterialDestroy);
+  model->materials[index] = material;
+  lovrRetain(material);
+  return true;
 }
 
 static bool lovrModelAnimateVertices(Model* model) {
@@ -8965,7 +9094,7 @@ static bool drawNode(Pass* pass, Model* model, uint32_t index, uint32_t instance
 
       DrawInfo draw = {
         .mode = part->mode == DRAW_POINT_LIST ? DRAW_POINTS : part->mode == DRAW_LINE_LIST ? DRAW_LINES : DRAW_TRIANGLES,
-        .material = model->materials && part->material != ~0u ? model->materials[part->material] : NULL,
+        .material = part->material != ~0u ? model->materials[part->material] : NULL,
         .transform = node->skin == ~0u ? globalTransform : NULL,
         .bounds = bounds,
         .vertex.buffer = model->vertexBuffer,
@@ -9042,7 +9171,7 @@ bool lovrPassDrawPart(Pass* pass, Model* model, uint32_t meshIndex, uint32_t par
 
     DrawInfo draw = {
       .mode = part->mode == DRAW_POINT_LIST ? DRAW_POINTS : part->mode == DRAW_LINE_LIST ? DRAW_LINES : DRAW_TRIANGLES,
-      .material = model->materials && part->material != ~0u ? model->materials[part->material] : NULL,
+      .material = part->material != ~0u ? model->materials[part->material] : NULL,
       .transform = transform, // TODO fix skinned mesh transforms?
       .bounds = part->bounds,
       .vertex.buffer = model->vertexBuffer,
@@ -9771,20 +9900,16 @@ static void trackTexture(Pass* pass, Texture* texture, gpu_phase phase, gpu_cach
 }
 
 static void trackMaterial(Pass* pass, Material* material) {
-  if (!material->hasWritableTexture) {
-    return;
-  }
-
   gpu_phase phase = GPU_PHASE_SHADER_VERTEX | GPU_PHASE_SHADER_FRAGMENT;
   gpu_cache cache = GPU_CACHE_TEXTURE;
 
-  trackTexture(pass, material->info.texture, phase, cache);
-  trackTexture(pass, material->info.glowTexture, phase, cache);
-  trackTexture(pass, material->info.metalnessTexture, phase, cache);
-  trackTexture(pass, material->info.roughnessTexture, phase, cache);
-  trackTexture(pass, material->info.clearcoatTexture, phase, cache);
-  trackTexture(pass, material->info.occlusionTexture, phase, cache);
-  trackTexture(pass, material->info.normalTexture, phase, cache);
+  trackTexture(pass, material->textures[TEXTURE_COLOR], phase, cache);
+  trackTexture(pass, material->textures[TEXTURE_GLOW], phase, cache);
+  trackTexture(pass, material->textures[TEXTURE_METALNESS], phase, cache);
+  trackTexture(pass, material->textures[TEXTURE_ROUGHNESS], phase, cache);
+  trackTexture(pass, material->textures[TEXTURE_CLEARCOAT], phase, cache);
+  trackTexture(pass, material->textures[TEXTURE_OCCLUSION], phase, cache);
+  trackTexture(pass, material->textures[TEXTURE_NORMAL], phase, cache);
 }
 
 static void trackRaytracer(Pass* pass, Raytracer* raytracer, gpu_phase phase, gpu_cache cache) {